+++ lib/pageview.dart
@@ -1,0 +1,6 @@
+/// Flutter widgets implementing custom page views.
+///
+/// To use, import `package:lit_ui_kit/buttons.dart`.
+library pageview;
+
+export 'src/widgets/pageview/lit_text_page_view.dart';