+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix overflow on `LitPrivacyPolicyScreen`.
 - Allow custom padding on `LitOnboardingScreen`.
+- Add `LitTextPageView` widget to display a list of cards with a tranition animation.
 
 ## 1.1.0